# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Meyora API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register_api_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/auth/switch-organization": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Switch Organization",
    "operationId": "switch_organization_api_auth_switch_organization_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SwitchOrganizationIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/auth/accept-invitation": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Accept Invitation",
    "operationId": "accept_invitation_api_auth_accept_invitation_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AcceptInvitationIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/api/auth/google/callback": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Google Callback",
    "operationId": "google_callback_api_auth_google_callback_get",
    "parameters": [
     {
      "name": "code",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Code"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "error",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Error"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AcceptInvitationIn": {
    "properties": {
     "token": {
      "type": "string",
      "maxLength": 200,
      "title": "Token"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "token"
    ],
    "title": "AcceptInvitationIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginIn": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginIn"
   },
...
   "MeOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     },
     "memberships": {
      "items": {
       "$ref": "#/components/schemas/MembershipOut"
      },
      "type": "array",
      "title": "Memberships"
     },
     "current_organization_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Organization Id"
     },
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Role"
       },
       {
        "type": "null"
       }
      ]
     },
     "permissions": {
      "items": {
       "$ref": "#/components/schemas/Perm"
      },
      "type": "array",
      "title": "Permissions"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "avatar_url",
     "memberships",
     "current_organization_id",
     "role",
     "permissions"
    ],
    "title": "MeOut"
   },
...
   "MembershipOut": {
    "properties": {
     "organization": {
      "$ref": "#/components/schemas/OrganizationOut"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     }
    },
    "type": "object",
    "required": [
     "organization",
     "role"
    ],
    "title": "MembershipOut"
   },
...
   "OrganizationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "default_currency": {
      "type": "string",
      "title": "Default Currency"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "slug",
     "default_currency"
    ],
    "title": "OrganizationOut"
   },
...
   "Perm": {
    "type": "string",
    "enum": [
     "crm:read",
     "crm:write",
     "crm:delete",
     "pipelines:manage",
     "members:read",
     "members:manage",
     "org:manage",
     "org:delete",
     "audit:read"
    ],
    "title": "Perm"
   },
...
   "RegisterIn": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 10,
      "title": "Password"
     },
     "full_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Full Name"
     },
     "organization_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Organization Name"
     },
     "invite_token": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Invite Token"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "email",
     "password",
     "full_name"
    ],
    "title": "RegisterIn"
   },
   "Role": {
    "type": "string",
    "enum": [
     "owner",
     "admin",
     "manager",
     "member"
    ],
    "title": "Role"
   },
...
   "SwitchOrganizationIn": {
    "properties": {
     "organization_id": {
      "type": "string",
      "format": "uuid",
      "title": "Organization Id"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "organization_id"
    ],
    "title": "SwitchOrganizationIn"
   },
...
   "TokenOut": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "me": {
      "$ref": "#/components/schemas/MeOut"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "me"
    ],
    "title": "TokenOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}